Add unit tests for item backstage passes
Item does not follow normal behaviour and increases in quality as sellIn
decreases. Quality increases by 2 when sellIn is 10 or less and by 3
when sellIn is 5 or less. When sellIn changes from 0 to -1 the quality
is dropped to 0.

Test coverage for the updateQuality function will move to green but not
100% due to repetition of logic in legacy code.