Create item class backstage pass
Specific item behaviour can be moved to its own class which inherits
from the item class. Create a backstage pass class with it's own
updateQuality function as the quality increases as sellIn decreases. The
increase value is conditional on the sellIn value.

The aged brie class can be used as a scaffold.

Add new feature and unit tests. Update existing unit tests for shop to
remove duplication.